% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_kegg.R
\name{.api_pull}
\alias{.api_pull}
\title{.api_pull}
\usage{
.api_pull(species, path = path)
}
\arguments{
\item{species}{character. The species to use in kegg data pull}

\item{path}{character. A character string describing the path to write out KEGG
API data sets. If not provided, defaults to current working directory.}
}
\value{
kegg_out: A named list of the data pulled from kegg api when the
function was run. This may be different if the function is run at
different times. For reproducible results, use text files generated
by function that include the date they were pulled.
}
\description{
This function connects to the KEGG API, downloads, and cleans
ncbi gene ID data, KEGG pathway descriptions, and species specific data.
Currently, this function supports Human, Mouse, and Rat. Files will be
written to the working directory unless otherwise specified by the user.
}
