% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\docType{data}
\name{kegg}
\alias{kegg}
\title{kegg}
\format{A \code{list} with three names items which are:
\describe{
\item{kegg_to_pathway}{List of kegg IDs mapped to pathway IDs }
\item{ncbi_to_kegg}{ List of ENTREZ gene IDs that map to kegg IDs}
\item{pathway_to_species}{ List of pathways IDs that map to rat pathways}
}}
\usage{
kegg
}
\description{
This is a \code{list} object that contains the output generated
from the \code{get_kegg} function. This object is
mostly meant for running examples and vignettes.
}
\keyword{datasets}
