% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFoldEnrichment.R
\name{plotFoldEnrichment}
\alias{plotFoldEnrichment}
\title{plotFoldEnrichment}
\usage{
plotFoldEnrichment(de_res, pval, N)
}
\arguments{
\item{de_res}{Dataframe. Generated using \code{\link{diffEnrich}}}

\item{pval}{Numeric. Threshold for filtering pathways based on adjusted pvalue in de_res}

\item{N}{Numeric. Number of top pathways to plot after filtering based on pval}
}
\value{
ggplot object. If the user calls \code{\link{plotFoldEnrich}} and
assigns it to an object (see example) then no plot will print in viewer,
but if \code{\link{plotFoldEnrich}} is called without being assigned to an
object the plot will print to the viewer. Users can edit the ggplot object
as they would any other ggplot object (e.g. add title, theme, etc.).
}
\description{
This function uses the results generated using \code{\link{diffEnrich}}
to generate a bar plot describing the fold enrichment of a set of given KEGG
pathways stratified by their enrichment in list 1 or list 2. Users can plot
all pathways based on the adjusted p-value threshold used in \code{\link{diffEnrich}}
and the top N pathways sorted by the adjusted p-value threshold used in \code{\link{diffEnrich}}.
\code{\link{plotFoldEnrich}} returns a ggplot2 object so users can add additional
customizations.
}
\examples{
\dontrun{
plot <- plotFoldEnrichment(de_res = diffEnrich, pval = 0.05, N = 5)
}
}
