% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffEnrich.R
\name{diffEnrich}
\alias{diffEnrich}
\title{diffEnrich}
\usage{
diffEnrich(list1_pe, list2_pe, method = "BH")
}
\arguments{
\item{list1_pe}{data.frame. Dataframe of enrichment results for genes of interest
generated from \code{\link{pathEnrich}}. See example for \code{\link{pathEnrich}}.}

\item{list2_pe}{data.frame. Dataframe of enrichment results for background genes
generated from \code{\link{pathEnrich}}. See example for \code{\link{pathEnrich}}.}

\item{method}{character. Character string telling \code{diffEnrich} which method to
use for multiple testing correction. Available methods are thos provided by
\code{\link{p.adjust}}, and the default is "BH", or False Discovery Rate (FDR).}
}
\value{
data.frame. Dataframe generated from merging pathEnrich dataframes with the following added columns:
                    Estimate: Estimated odds ration calculated from Fisher's Exact test
                    P_value: Unadjusted p_value from Fisher's Exact test
}
\description{
This function takes the objects generated from \code{\link{pathEnrich}}.
If performing a dfferential enrichment analysis, the user will have 2 objects. There
will be one for the genes of interest and one for the background (see example for \code{\link{pathEnrich}}).
This function then uses a Fisher's Exact test to iddentify differentially enriched
pathways between the terms enriched in the gene-of-interest list and the pathways enriched
in the background. \code{diffEnrich} returns a dataframe containing differentially enriched
pathways with their associated estimated odds ratio, unadjusted p-value, and fdr adjusted
p-value.
}
\examples{
## Generate individual enrichment reults
list1_pe <- pathEnrich(gk_obj = kegg, gene_list = geneLists$list1)
list2_pe <- pathEnrich(gk_obj = kegg, gene_list = geneLists$list2)

## Perform differential enrichment
dif_enrich <- diffEnrich(list1_pe = list1_pe, list2_pe = list2_pe, method = 'none')

}
