% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffEnrich.R
\name{diffEnrich}
\alias{diffEnrich}
\title{diffEnrich}
\usage{
diffEnrich(list1_pe, list2_pe, method = "BH")
}
\arguments{
\item{list1_pe}{data.frame. Dataframe of enrichment results for genes of interest
generated from \code{\link{pathEnrich}}. See example for \code{\link{pathEnrich}}.}

\item{list2_pe}{data.frame. Dataframe of enrichment results for background genes
generated from \code{\link{pathEnrich}}. See example for \code{\link{pathEnrich}}.}

\item{method}{character. Character string telling \code{diffEnrich} which method to
use for multiple testing correction. Available methods are thos provided by
\code{\link{p.adjust}}, and the default is "BH", or False Discovery Rate (FDR).}
}
\value{
res: An object of class data.frame that summarizes the results
of the differential enrichment and contains the following variables:

\describe{
  \item{KEGG_PATHWAY_ID}{KEGG Pathway Identifier}
  \item{KEGG_PATHWAY_description}{Description of KEGG Pathway (provided by KEGG)}
  \item{KEGG_PATHWAY_cnt}{Number of Genes in KEGG Pathway}
  \item{KEGG_DATABASE_cnt}{Number of Genes in KEGG Database}
  \item{KEGG_PATHWAY_in_list1}{Number of Genes from gene list 1 in KEGG Pathway}
  \item{KEGG_DATABASE_in_list1}{Number of Genes from gene list 1 in KEGG Database}
  \item{expected_list1}{Expected number of genes from list 1 to be in KEGG pathway by chance (i.e., not enriched)}
  \item{enrich_p_list1}{P-value for enrichment of list 1 genes related to KEGG pathway}
  \item{fdr_list1}{False Discovery Rate (Benjamini and Hochberg) of enrich_p_list1 to account for multiple testing across KEGG pathways}
  \item{KEGG_PATHWAY_in_list2}{Number of Genes from gene list 2 in KEGG Pathway}
  \item{KEGG_DATABASE_in_list2}{Number of Genes from gene list 2 in KEGG Database}
  \item{expected_list2}{Expected number of genes from list 2 to be in KEGG pathway by chance (i.e., not enriched)}
  \item{enrich_p_list2}{P-value for enrichment of list 2 genes related to KEGG pathway}
  \item{fdr_list2}{False Discovery Rate (Benjamini and Hochberg) of enrich_p_list2 to account for multiple testing across KEGG pathways}
  \item{odd_ratio}{Odds of a gene from list 2 being from this KEGG pathway / Odds of a gene from list 1 being from this KEGG pathway}
  \item{diff_enrich_p}{P-value for differential enrichment of this KEGG pathway between list 1 and list 2}
  \item{diff_enrich_adjusted}{Multiple testing adjustment of diff_enrich_p (default = False Discovery Rate (Benjamini and Hochberg))}
}
}
\description{
This function takes the objects generated from \code{\link{pathEnrich}}.
If performing a dfferential enrichment analysis, the user will have 2 objects. There
will be one for the genes of interest and one for the background (see example for \code{\link{pathEnrich}}).
This function then uses a Fisher's Exact test to iddentify differentially enriched
pathways between the terms enriched in the gene-of-interest list and the pathways enriched
in the background. \code{diffEnrich} returns a dataframe containing differentially enriched
pathways with their associated estimated odds ratio, unadjusted p-value, and fdr adjusted
p-value.
}
\examples{
## Generate individual enrichment reults
list1_pe <- pathEnrich(gk_obj = kegg, gene_list = geneLists$list1)
list2_pe <- pathEnrich(gk_obj = kegg, gene_list = geneLists$list2)

## Perform differential enrichment
dif_enrich <- diffEnrich(list1_pe = list1_pe, list2_pe = list2_pe, method = 'none')

}
