% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffEnrich.R
\name{.combineEnrich}
\alias{.combineEnrich}
\title{.combineEnrich}
\usage{
.combineEnrich(sig_pe, bkg_pe)
}
\arguments{
\item{sig_pe}{data.frame. Dataframe of enrichment results for genes of interest
generated from \code{\link{pathEnrich}}. See example for \code{\link{pathEnrich}}.}

\item{bkg_pe}{data.frame. Dataframe of enrichment results for background genes
generated from \code{\link{pathEnrich}}. See example for \code{\link{pathEnrich}}.}
}
\value{
combined_enrich: An object of class data.frame that is the result of merging
\code{sig_pe} and \code{bkg_pe}, using the default joining implemented in the base
\code{\link{merge}} function.
}
\description{
This is a helper function for \code{diffEnrich}. This function takes the objects generated from \code{\link{pathEnrich}}.
If performing a dfferential enrichment analysis, the user will have 2 objects. There
will be one for the genes of interest and one for the background (see example for \code{\link{pathEnrich}}).
This function then merges the two dataframes using the following columns that should be present
in both objects (\code{by = c("V1", "V2", "KEGG_cnt", "numTested")}). This merged dataframe
will be used as the input for the differential enrichment function.
}
