% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\docType{data}
\name{geneLists}
\alias{geneLists}
\title{geneLists}
\format{A \code{list} with two names items which are:
\describe{
\item{background}{List of ENTREZ gene IDs that will considered background }
\item{sigGenes}{ List of ENTREZ gene IDs that were significant}
}}
\usage{
geneLists
}
\description{
This is a \code{list} object that contains the list background
genes and significant genes used in pathway enrichment. This object is
mostly meant for running examples and vignettes. The data provided is
for the rat, and is loaded from org.Rn.eg.db version 3.7.0.
}
\keyword{datasets}
