% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_kegg.R
\name{get_kegg}
\alias{get_kegg}
\title{get_kegg}
\usage{
get_kegg(species)
}
\arguments{
\item{species}{character. The species to use in kegg data pull}
}
\value{
kegg_out: A named list of the data pulled from kegg api when the
function was run. This may be different if the function is run at
different times. For reproducible results, use text files generated
by function that include the date they were pulled.
kegg_out: "ncbi_to_kegg" = ncbi_to_kegg dataframe,
          "kegg_to_pathway" = kegg_to_pathway dataframe,
          "pathway_to_species" = pathway_to_species dataframe
}
\description{
This function connects to the KEGG API, downloads, and cleans
ncbi gene ID data, KEGG pathway descriptions, and species specific data.
Currently, this function supports Human, Mouse, and Rat. Files will be
written to the working directory.
}
\examples{
kegg <- get_kegg(species = "human")
\dontrun{
kegg <- get_kegg(species = "mouse")
}

}
