% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathEnrich.R
\name{pathEnrich}
\alias{pathEnrich}
\title{pathEnrich}
\usage{
pathEnrich(gk_obj, gene_list)
}
\arguments{
\item{gk_obj}{list. Object genrated from \code{get_kegg}, or a list containing the
output generated from a past \code{get_kegg} call. Names of the list must match those
defined in \code{get_kegg}. If the user wishes to use an older version of data
generated by \code{get_kegg}, they must first load that data and put it in a named
list that matches the names given in the list generated by \code{get_kegg}.}

\item{gene_list}{Vector. Vector of NCBI (ENTREZ) geneIDs.}
}
\value{
enrich_table: An object of class data.frame that summarizes the results
of the pathway analysis and contains the following variables:

\describe{
  \item{KEGG_PATHWAY_ID}{KEGG Pathway Identifier}
  \item{KEGG_PATHWAY_description}{Description of KEGG Pathway (provided by KEGG)}
  \item{KEGG_PATHWAY_cnt}{Number of Genes in KEGG Pathway}
  \item{KEGG_PATHWAY_in_list}{Number of Genes from gene list in KEGG Pathway}
  \item{KEGG_DATABASE_cnt}{Number of Genes in KEGG Database}
  \item{KEGG_DATABASE_in_list}{Number of Genes from gene list in KEGG Database}
  \item{expected}{Expected number of genes from list to be in KEGG pathway by chance (i.e., not enriched)}
  \item{enrich_p}{P-value for enrichment of list genes related to KEGG pathway}
  \item{fdr}{False Discovery Rate (Benjamini and Hochberg) to account for multiple testing across KEGG pathways}
}
}
\description{
This function takes the list generated in \code{\link{get_kegg}} as well as a vector
of NCBI (ENTREZ) geneIDs, and identifies significantly enriched KEGG pathways using
a Fisher's Exact Test. Unadjusted p-values as well as FDR corrected p-values are
calculated.
}
\details{
This function may not always use the complete list of genes provided by the user.
Specifically, it will only use the genes from the list provided that are also in
the most current species list pulled from the KEGG REST API, or from the older data KEGG
loaded by the user.
}
\examples{

list1_pe <- pathEnrich(gk_obj = kegg, gene_list = geneLists$list1)
\dontrun{
list2_pe <- pathEnrich(gk_obj = kegg, gene_list = geneLists$list2)
}

}
